backend/requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
pydantic
